Improve sync, fixed bugs, rename files, add rebuild, more robust loading

+++ inc/ti/events.h
@@ -16,7 +16,7 @@ typedef struct ti_events_s ti_events_t;
 int ti_events_create(void);
 int ti_events_start(void);
 void ti_events_stop(void);
-int ti_events_trigger_loop(void);
+ssize_t ti_events_trigger_loop(void);
 int ti_events_on_event(ti_node_t * from_node, ti_pkg_t * pkg);
 int ti_events_create_new_event(ti_query_t * query, ex_t * e);
 int ti_events_add_event(ti_node_t * node, ti_epkg_t * epkg);
